Cleanup CXF
- Renames CXP to CXF
- Move login logic to separate file
