--- crates/bitwarden-exporters/src/cxp/mod.rs
@@ -1,8 +1,0 @@
-mod error;
-pub use error::CxpError;
-
-mod export;
-pub(crate) use export::build_cxf;
-pub use export::Account;
-mod import;
-pub(crate) use import::parse_cxf;
