+++ crates/bitwarden-exporters/src/error.rs
@@ -10,7 +10,7 @@ pub enum ExportError {
     #[error("CSV error: {0}")]
     Csv(#[from] crate::csv::CsvError),
     #[error("Credential Exchange error: {0}")]
-    CxpError(#[from] crate::cxp::CxpError),
+    CxpError(#[from] crate::cxf::CxfError),
     #[error("JSON error: {0}")]
     Json(#[from] crate::json::JsonError),
     #[error("Encrypted JSON error: {0}")]
